Changing label types from int64 to int32.
For classification datasets that already have unit tests, the type of label series is changed from default (int64) to int32 (that is expected by most ML libraries). For datasets that do not have unit tests yet, this needs to be done in future commits.

+++ training/xtime/datasets/_gesture_phase.py
@@ -76,7 +76,7 @@ class GesturePhaseSegmentationBuilder(DatasetBuilder):
         assert isinstance(y, pd.Series), f"Expecting y to be of type pd.Series (type = {type(y)})."
 
         # Encode labels
-        y = pd.Series(LabelEncoder().fit_transform(y), index=y.index, name=y.name)
+        y = pd.Series(LabelEncoder().fit_transform(y), index=y.index, name=y.name).astype("int32")
 
         # Remove instances with missing values
         # No need to remove instances with missing values - no such instances.
